Package renaming after changing mastery check package
