+++ src/main/java/seedu/jarvis/logic/parser/ConsultationMasteryCheckCommandParser.java
@@ -11,7 +11,7 @@ import java.time.format.DateTimeParseException;
 import seedu.jarvis.logic.commands.add.AddCommand;
 import seedu.jarvis.logic.parser.exceptions.ParseException;
 import seedu.jarvis.model.consultation.Consultation;
-import seedu.jarvis.model.consultation.MasteryCheck;
+import seedu.jarvis.model.masteryCheck.MasteryCheck;
 
 /**
  * Parses input add consultation or mastery check arguments and creates a new specified consultation or mc object
