+++ src/main/java/seedu/jarvis/model/ReadOnlyAddressBook.java
@@ -3,7 +3,7 @@ package seedu.jarvis.model;
 import javafx.beans.property.StringProperty;
 import javafx.collections.ObservableList;
 import seedu.jarvis.model.consultation.Consultation;
-import seedu.jarvis.model.consultation.MasteryCheck;
+import seedu.jarvis.model.masteryCheck.MasteryCheck;
 import seedu.jarvis.model.mission.Mission;
 import seedu.jarvis.model.quest.Quest;
 import seedu.jarvis.model.student.Student;
